server/udp: implement zombie reaper thread
Introducing zombie reaper thread (zr thread).

In order to efficiently reap zombie sessions, we implement zombie
reaper thread which will send request sync packet to authenticated
clients if it reaches the timeout.

It will also close() and free() the sessions those are assumed dead by
the given timeout. The timeout for the authenticated session and
unauthenticated session are different.

For authenticated session, it's `UDP_SESS_TIMEOUT * 2` seconds.
For unauthenticated session, it's `UDP_SESS_TIMEOUT` seconds.

If an authenticated session reaches `UDP_SESS_TIMEOUT`, the zr thread
will send a sync request packet (in hope that the client will reply).
If the client replies, the timeout counter for it will be reset.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>