server/udp: zombie-reaper: fix potential data race with epoll threads
Calling `close_udp_session()` from zombie reaper thread is the wrong
thing to do. Because it uses `struct epl_thread` buffer, which is
being used by other threads. This leads to data race because there
is no locking mechanism that protects those buffers.

Fix for this is that, allocate separated buffer for zombie reaper
thread so it does not clobber other threads' buffer.

Fixes: f65243c69a7d6fa0233533400ce709fcf5ecd5ff ("server/udp: implement zombie reaper thread")
Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>