server/udp: implement update timeout counter in the event loop
Add a new field `loop_c` to the `struct udp_sess`, it's used for
periodically updating the timeout counter (`last_act`).

How does it work?
Whenever the event loop calls `recvfrom()` and it has corresponding
online session in the map, it increments the `loop_c`. After that,
it checks the value of `loop_c`, if it is divisible by 8, then update
the timeout counter by calling `udp_sess_tv_update()`.

The reason of using `loop_c` is to reduce the number of
`gettimeofday()` syscalls being invoked.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>